+++ clients/python/CHANGELOG.md
@@ -14,4 +14,5 @@ and **Merged pull requests**. Critical items to know are:
 The versions coincide with releases on pip. Only major versions will be released as tags on Github.
 
 ## [0.0.x](https://github.com/flux-framework/flux-restful-api/tree/main) (0.0.x)
+ - Parse envars from the command line, add descriptions for submit (0.0.1)
  - Project skeleton release (0.0.0)